refactor: extract @mulmocast/script-utils from preprocessor
Split pure data transformation utilities (filter, variant, profiles,
process, context-builder) into a new lightweight package that has no
GraphAI dependency. Consumers import directly from @mulmocast/script-utils
instead of through re-exports.
