--- packages/mulmocast-preprocessor/src/context.ts
@@ -1,8 +1,0 @@
-// Lightweight entry point for context-building utilities (no GraphAI dependency)
-export {
-  buildScriptContent,
-  buildBeatContent,
-  buildScriptMetaContent,
-  scriptToViewerData,
-  DEFAULT_INTERACTIVE_SYSTEM_PROMPT,
-} from "./core/ai/context-builder.js";
