+++ packages/mulmocast-script-utils/src/variant.ts
